#7 Separate encode and decode methods from QuanvolutionalFilter
